# open_api/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace Product API
  description: API для управления товарами в маркетплейсе с поддержкой CRUD операций
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
...
paths:
  /products:
    post:
      tags:
        - Products
      summary: Создание товара
      description: Создает новый товар в системе
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Products
      summary: Список товаров с пагинацией и фильтрацией
      description: Возвращает список товаров с поддержкой пагинации и фильтрации по статусу и категории
      operationId: getProducts
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          description: Фильтр по статусу товара
          required: false
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: category
          in: query
          description: Фильтр по категории (точное совпадение)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Список товаров успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          description: Некорректные параметры запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders:
    post:
      tags:
        - Orders
      summary: Создание заказа
      description: Создает новый заказ с проверкой всех бизнес-правил
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Товар или промокод не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Конфликт (недостаточно товара, активный заказ и т.д.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Ошибка промокода
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит частоты запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}:
    put:
      tags:
        - Orders
      summary: Обновление заказа
      description: Обновляет состав заказа (только в статусе CREATED)
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          description: ID заказа
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Заказ успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый статус или конфликт
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Превышен лимит частоты запросов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /orders/{id}/cancel:
    post:
      tags:
        - Orders
      summary: Отмена заказа
      description: Отменяет заказ (только из статусов CREATED, PAYMENT_PENDING)
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          description: ID заказа
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заказ успешно отменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          description: Заказ принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый статус для отмены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Получение товара по ID
      description: Возвращает информацию о товаре по его идентификатору
      operationId: getProductById
      parameters:
        - name: id
          in: path
          description: Уникальный идентификатор товара
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Товар успешно найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Products
      summary: Обновление товара
      description: Обновляет информацию о существующем товаре
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          description: Уникальный идентификатор товара
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Товар успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Некорректные данные запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Products
      summary: Мягкое удаление товара
      description: Переводит товар в статус ARCHIVED (мягкое удаление)
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          description: Уникальный идентификатор товара
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Товар успешно архивирован
        '404':
          description: Товар не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ARCHIVED
      description: |
        Статус товара:
        * `ACTIVE` - Активный товар, доступен для покупки
        * `INACTIVE` - Неактивный товар, временно недоступен
        * `ARCHIVED` - Архивный товар, удален из каталога
    ProductCreate:
      type: object
      required:
        - name
        - price
        - category
        - stock
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          nullable: true
          maxLength: 4000
          description: Описание товара (необязательное поле)
        price:
          type: number
          format: double
          minimum: 0.01
          description: Цена товара (должна быть больше 0)
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        stock:
          type: integer
          minimum: 0
          description: Количество товара на складе (должно быть >= 0)
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Название товара
        description:
          type: string
          nullable: true
          maxLength: 4000
          description: Описание товара
        price:
          type: number
          format: double
          minimum: 0.01
          description: Цена товара (должна быть больше 0)
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Категория товара
        stock:
          type: integer
          minimum: 0
          description: Количество товара на складе (должно быть >= 0)
        status:
          $ref: '#/components/schemas/ProductStatus'
    OrderCreate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          minItems: 1
          maxItems: 50
          description: Список позиций заказа
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true
          description: Промокод
    OrderUpdate:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          minItems: 1
          maxItems: 50
          description: Новый список позиций заказа
    OrderItemRequest:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: string
          format: uuid
          description: ID товара
        quantity:
          type: integer
          minimum: 1
          maximum: 999
          description: Количество товара
...
    OrderResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        total_amount:
          type: number
          format: double
        discount_amount:
          type: number
          format: double
        promo_code_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - category
        - stock
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор товара
        name:
          type: string
          description: Название товара
        description:
          type: string
          nullable: true
          description: Описание товара
        price:
          type: number
          format: double
          description: Цена товара
        category:
          type: string
          description: Категория товара
        stock:
          type: integer
          description: Количество товара на складе
        status:
          $ref: '#/components/schemas/ProductStatus'
        created_at:
          type: string
          format: date-time
          description: Дата и время создания товара (только в response)
        updated_at:
          type: string
          format: date-time
          description: Дата и время последнего обновления товара (только в response)
    ProductListResponse:
      type: object
      required:
        - content
        - totalElements
        - page
        - size
      properties:
        content:
          type: array
          description: Список товаров на текущей странице
          items:
            $ref: '#/components/schemas/ProductResponse'
        totalElements:
          type: integer
          description: Общее количество товаров
        page:
          type: integer
          description: Номер текущей страницы (начиная с 0)
        size:
          type: integer
          description: Размер страницы
        totalPages:
          type: integer
          description: Общее количество страниц
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          type: string
          description: |
            Машиночитаемый код ошибки:
            * `PRODUCT_NOT_FOUND` - Товар не найден по ID (404)
            * `PRODUCT_INACTIVE` - Попытка заказать неактивный товар (409)
            * `ORDER_NOT_FOUND` - Заказ не найден по ID (404)
            * `ORDER_LIMIT_EXCEEDED` - Превышен лимит частоты создания/обновления заказа (429)
            * `ORDER_HAS_ACTIVE` - У пользователя уже есть активный заказ (409)
            * `INVALID_STATE_TRANSITION` - Недопустимый переход состояния заказа (409)
            * `INSUFFICIENT_STOCK` - Недостаточно товара на складе (409)
            * `PROMO_CODE_INVALID` - Промокод не найден, истёк, исчерпан или неактивен (422)
            * `PROMO_CODE_MIN_AMOUNT` - Сумма заказа ниже минимальной для промокода (422)
            * `ORDER_OWNERSHIP_VIOLATION` - Заказ принадлежит другому пользователю (403)
            * `VALIDATION_ERROR` - Ошибка валидации входных данных (400)
        message:
          type: string
          description: Человекочитаемое описание
        details:
          type: object
          description: Дополнительные данные (опционально)
          additionalProperties: true